requirements.txt
pillow
matplotlib
glob
torch
tqdm
shutil